feat: merge modules into unified commands
